Change query style from top down to bottom up from school assessments
The queries for GP contest were done in a top-down style where, the
blocks in a district are identified first, then for each block, the
clusters are identified and for each cluster schools are
identified. This means that the schools that does not participate in
GP contest are also considered. To fix this, a bottom-up approach is
used, where the assessment data is filtered such that it does not
consider schools that does not participate in GP contest.